+++ src/config/settings.production.js
@@ -2,5 +2,5 @@
 // https://create-react-app.dev/docs/adding-custom-environment-variables/
 export default {
   environment: "production",
-  API_BASE_URL: process.env.REACT_APP_API_BASE_URL
+  API_BASE_URL: process.env.REACT_APP_API_BASE_URL,
 };