+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-### Added 
+### Added
+- Allow kudometer names to be changed [PR](https://github.com/kabisa/kudos-frontend/pull/75), [Pivotal issue](https://www.pivotaltracker.com/story/show/161506981)
+- Allow the active kudometer to be changed [PR](https://github.com/kabisa/kudos-frontend/pull/75), [Pivotal issue](https://www.pivotaltracker.com/story/show/163294940)
 - Button to show all guidelines
-
+- Integrations section to team management page [PR](https://github.com/kabisa/kudos-frontend/pull/72), [Pivotal issue](https://www.pivotaltracker.com/story/show/172309004)
+- Slack register section to user profile page [PR](https://github.com/kabisa/kudos-frontend/pull/72), [Pivotal issue](https://www.pivotaltracker.com/story/show/172308999)
 ### Fixed
 - Receiver autocomplete is now cleared after selecting an entry. [PR](https://github.com/kabisa/kudos-frontend/pull/69), [Pivotal issue](https://www.pivotaltracker.com/story/show/162732488) 
 - Refreshing the team management page will now stay on the active section. [PR](https://github.com/kabisa/kudos-frontend/pull/69)